add account connection messages
