+++ packages/bridge-ui-v2/src/components/AccountConnectionToast/index.ts
@@ -1,0 +1,1 @@
+export { default as AccountConnectionToast } from './AccountConnectionToast.svelte';
