+++ packages/bridge-ui-v2/src/libs/emitter/emitter.ts
@@ -1,0 +1,3 @@
+import { EventEmitter } from 'events';
+
+export const emitter = new EventEmitter();
