+++ packages/bridge-ui-v2/src/libs/emitter/index.ts
@@ -1,0 +1,1 @@
+export { emitter } from './emitter';
