+++ packages/bridge-ui-v2/src/routes/+layout.svelte
@@ -4,23 +4,23 @@
 
   import { onDestroy, onMount } from 'svelte';
 
+  import { AccountConnectionToast } from '$components/AccountConnectionToast';
   import { Header } from '$components/Header';
   import { NotificationToast } from '$components/NotificationToast';
   import { SideNavigation } from '$components/SideNavigation';
   import { startWatching, stopWatching } from '$libs/wagmi';
-  import { account } from '$stores/account';
-
-  $: connected = Boolean($account?.isConnected);
 
   onMount(startWatching);
   onDestroy(stopWatching);
 </script>
 
 <SideNavigation>
-  <Header {connected} />
+  <Header />
   <main>
     <slot />
   </main>
 </SideNavigation>
 
 <NotificationToast />
+
+<AccountConnectionToast />
